refactor: add package for enumerated types
