+++ src/main/kotlin/ecommerce/entity/Member.kt
@@ -1,5 +1,6 @@
 package ecommerce.entity
 
+import ecommerce.entity.enumerated.Role
 import jakarta.persistence.Column
 import jakarta.persistence.Entity
 import jakarta.persistence.EnumType
